Common subsequence intervals extraction algorithm now prioritizes delete steps and got a fix. Some reorganising of verbs.py.
The algorithms now prioritizes delete steps over insert and edit steps in case of a draw regarding cost reduction. Current evidence suggests that this is better as it splits the words "liegen" and "gelegen" into |  |l|i|egen| and |ge|l| |egen| instead of | li|egen| and |gel|egen|, however there might be other cases where this is not true, so further evaluation is needed. A possibility would also be to change the costs of edit, insert and delete step in the LCS matrix computation to tweak the result to the desired.