Introduced module input_parsing with base class WordProcessor and implementations HangeulComposer and CombinedProcessor.
WordProcessor is an interface for all implementations of processing on the direct input words.
HangeulComposer is a WordProcessor which converts all composed Hangeul syllables in the input (leaving all other characters untouched) and vice-versa in the output.
CombinedProcessor allows the subsequent execution of multiple WordProcessor instances. It accepts a list of instances and executes them one by one according to the lists order when processing input. When processing output, the order is reveresed.